Windows build falg added.
